2018/11/4 群組圖片身分限制

+++ app/Services/ChatService.php
@@ -60,24 +60,32 @@ class ChatService
         return $dataList;
     }
 
-    public function updateChatProfilePic(\Illuminate\Http\UploadedFile $file, $chat_id)
+    public function updateChatProfilePic(\Illuminate\Http\UploadedFile $file, $chat_id, $account)
     {
-        $newFileName = date("YmdHis", time()) . '___' . rand(1000, 9999) . '___' . $file->getClientOriginalName();
-        if (strlen($newFileName) > 200)
-            return '0';
-        $image = Image::make($file);
-        $image->resize(350, null, function ($constraint) {
-            $constraint->aspectRatio();
-            $constraint->upsize();
-        })->save('images/ChatProfilePic/' . $newFileName);
-        $Chat = ChatEloquent::find($chat_id);
-        if ($Chat->profile_pic) {
-            if (file_exists('images/ChatProfilePic/' . $Chat->profile_pic)) {
-                unlink('images/ChatProfilePic/' . $Chat->profile_pic);
+        $CMcheck = ChatMemberEloquent::where('account', $account)
+            ->where('chat_id', $chat_id)
+            ->where('status', 2)->first();
+        if ($CMcheck) {
+            $newFileName = date("YmdHis", time()) . '___' . rand(1000, 9999) . '___' . $file->getClientOriginalName();
+            if (strlen($newFileName) > 200)
+                return '0';
+            $image = Image::make($file);
+            $image->resize(350, null, function ($constraint) {
+                $constraint->aspectRatio();
+                $constraint->upsize();
+            })->save('images/ChatProfilePic/' . $newFileName);
+            $Chat = ChatEloquent::find($chat_id);
+            if ($Chat->profile_pic) {
+                if (file_exists('images/ChatProfilePic/' . $Chat->profile_pic)) {
+                    unlink('images/ChatProfilePic/' . $Chat->profile_pic);
+                }
             }
+            $Chat->profile_pic = $newFileName;
+            $Chat->save();
+            return $newFileName;
+        } else {
+            return '您並非該群組成員';
         }
-        $Chat->profile_pic = $newFileName;
-        $Chat->save();
-        return $newFileName;
+
     }
 }
